mtd: spi: renesas: Clear HS bit during hardware initialization
According to the datasheet, HS bit should be specified to 1 when using
DMA transfer. As DMA transfer is not supported, it should be cleared to 0

Previously, the driver relies on HS bit setting in prior firmware but
this is not always the case.

Fix this by ensuring the bit is cleared during hardware initialization.

Based on Linux commit 5192481f908e, and use clrsetbits_le32 to simplify
the register modification.

Signed-off-by: Hai Pham <hai.pham.ud@renesas.com>
Signed-off-by: Khanh Le <khanh.le.xr@renesas.com>