pinctrl: renesas: r8a779g0: Update PFC driver for Ethernet-AVB channel 0/1
According to the latest hardware specification update, the pin function
configuration for Ethernet-AVB channel 0/1 has been revised. This patch
updates the PFC driver to align with the new information from the HW
manual.
These changes do not affect the Ethernet-AVB driver or the operation of
Ethernet-AVB on R-Car V4H.

Signed-off-by: Tam Nguyen <tam.nguyen.xa@renesas.com>